+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [10.0.0] - 2020-03-20
 ### Added
  - Added struct for representing Windows device paths
  - Added IWindowsDevice interface for devices with a windows device path